Allow access to settings through a hash interface
This allows using `#[]` and `#[]=` to get/set settings in the
configuration class.

The idea is that, maybe, it can be useful in order to use the
configuration class as a container for dry-auto_inject (which expects
`#[]` interface) for user configured dependencies.

Also, it facilitates testing when the configuration class is resolved
from a container, because it allows to be easily interchanged with a
hash. Used in this way (but not if the configuration class dependency is
 hardcoded) it could be a replacement for #19