Add cloneable option to setting
With this, the author of the configurable object can specify explicitly whether given setting values should be cloned.

This is useful when providing a custom default value with its own “rich” configuration interface, something different to the usual Array, Hash, and Set “primitives,” which were previously the only values that could be cloned.