Deprecate constructor as a block
So far, we have been accepting both a block and the `constructor:`
keyword argument for providing a constructor. The former will not be
available in 1.0, so we're deprecating it now.

We also prepare the test suite for the change and make naming consistent
using "constructor" where we were using "pre-processor."