Fix test suite after change in deprecation's uplevel value in dry-core
After the legit change made at dry-rb/dry-core#56, the test suite here
was failing because of a wrong assumption made before: our tests are
calling from the DSL object itself in contrast to the class extending
it. I.e., we can't rely on the default value of the `uplevel` parameter
as we're skipping one stack.

References #115