Rename `_settings` class method to `settings`
This gives us a "neater" name for the second (after `config`) main way of interacting with configurable classes. It's also better match with the `setting` method we use for defining new settings in the first place.

This method was previously `_settings` because we had already had the `settings` name taken by the method that returns a set of setting names only. However, these names can be just as easily accessed via the `#name` attribute on individual settings themselves, and reducing our overall surface area here will help make the settings side of our API easier to understand and use.